refactor: split handoff command family
